feat: enhance BulkActionsBar with dynamic assignee and more actions menus, improving bulk event management functionality
